+++ test.sh
@@ -22,6 +22,7 @@ swiftc \
   "$ROOT_DIR/tests/swift/SetupCoreUtilitiesTests.swift" \
   "$ROOT_DIR/tests/swift/RedistInstallerTests.swift" \
   "$ROOT_DIR/tests/swift/SetupStatusToneTests.swift" \
+  "$ROOT_DIR/tests/swift/USVFSUpdaterTests.swift" \
   "$ROOT_DIR/tests/swift/main.swift" \
   -o "$SWIFT_TEST_BINARY"
 "$SWIFT_TEST_BINARY"